Fix node build fail
