+++ node/Cargo.toml
@@ -370,9 +370,10 @@ git = "https://github.com/peaqnetwork/cumulus"
 branch = "peaq-polkadot-v0.9.29"
 
 [dependencies.peaq-pallet-did-rpc]
-git = 'https://github.com/peaqnetwork/peaq-pallet-did.git'
-branch = 'dev'
-version = '0.0.1'
+path = '../../peaq-pallet-did/rpc'
+# git = 'https://github.com/peaqnetwork/peaq-pallet-did.git'
+# branch = 'peaq-polkadot-v0.9.29'
+# version = '0.0.2'
 
 
 # XCM
